requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Authlib==1.2.0
autopep8==2.0.2
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
billiard==4.2.0
blis==0.7.11
cachelib==0.10.2
catalogue==2.0.10
celery==5.3.6
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
cloudpathlib==0.15.1
colorama==0.4.6
comm==0.1.3
confection==0.1.3
cryptography==39.0.2
cycler==0.11.0
cymem==2.0.8
dataclasses-json==0.5.7
DateTime==5.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.0.7
descartes==1.1.0
docutils==0.20.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
exceptiongroup==1.1.0
executing==2.0.0
Fiona==1.9.2
flasgger==0.9.7.1
Flask==2.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Dance==6.2.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.0.3
Flask-SSLify==0.1.5
fonttools==4.29.1
frozenlist==1.4.0
geographiclib==2.0
geopandas==0.12.2
geopy==2.3.0
gpt-engineer==0.2.5
graphql-core==3.2.3
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
icecream==2.1.3
idna==3.4
importlib-metadata==7.0.0
influxdb==5.3.1
install==1.3.5
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter_client==8.2.0
jupyter_core==5.3.0
keyring==24.3.0
kiwisolver==1.3.2
kombu==5.3.4
langchain==0.0.353
langchain-community==0.0.7
langchain-core==0.1.4
langcodes==3.3.0
langsmith==0.0.75
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
monotonic==1.6
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
munch==2.5.0
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
netaddr==0.8.0
nh3==0.2.15
numpy==1.22.2
oauthlib==2.1.0
openai==0.28.0
outcome==1.2.0
packaging==23.2
pandas==1.4.1
parso==0.8.3
pathy==0.10.2
pickleshare==0.7.5
Pillow==9.0.1
pkginfo==1.9.6
platformdirs==3.5.1
plotly==5.18.0
preshed==3.0.9
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyGithub==1.59.1
Pygments==2.16.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.7
pypi-express-pub==0.0.5
pyproj==3.5.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.0
readme-renderer==42.0
redis==4.5.5
referencing==0.30.0
regex==2023.12.25
requests==2.28.2
requests-oauthlib==1.1.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
rpds-py==0.9.2
rudder-sdk-python==2.0.2
schedule==1.2.1
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.8.2
shapely==2.0.1
singledispatch==4.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spotipy==2.23.0
SQLAlchemy==2.0.5.post1
srsly==2.4.8
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
thinc==8.2.1
threadpoolctl==3.1.0
tiktoken==0.5.2
tomli==2.0.1
tornado==6.3.2
tqdm==4.62.3
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
twine==4.0.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.14
URLObject==2.4.3
vine==5.1.0
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.2
Werkzeug==2.2.2
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
zope.interface==5.5.2
